fix(redis,stream): address Copilot review
Copilot 3 review comments — comment 1 (DEL >100k regression) was
already resolved in 96ca38d3; the remaining two:

- xlen: replace conn.WriteInt(int(meta.Length)) with
  conn.WriteInt64(meta.Length). meta.Length is int64 and truncating
  to Go's int would overflow on 32-bit targets or any stream whose
  length legitimately exceeds MaxInt32. The legacy fallback is
  likewise updated for consistency (int64(len(entries))).
- marshalStreamEntry doc: the previous comment claimed only Fields
  were serialized, which disagrees with the implementation (Id is
  also written). Updated the comment to explain the intentional
  duplication — unmarshalStreamEntry returns a fully-formed entry
  without needing the caller to parse the storage key back.

+++ adapter/redis_compat_commands.go
@@ -4428,7 +4428,7 @@ func (r *RedisServer) xlen(conn redcon.Conn, cmd redcon.Command) {
 		return
 	}
 	if found {
-		conn.WriteInt(int(meta.Length))
+		conn.WriteInt64(meta.Length)
 		return
 	}
 	stream, err := r.loadStreamAt(context.Background(), cmd.Args[1], readTS)
@@ -4436,7 +4436,7 @@ func (r *RedisServer) xlen(conn redcon.Conn, cmd redcon.Command) {
 		conn.WriteError(err.Error())
 		return
 	}
-	conn.WriteInt(len(stream.Entries))
+	conn.WriteInt64(int64(len(stream.Entries)))
 }
 
 func parseRangeStreamCount(args [][]byte) (int, error) {
